makes parent pointer an option type
Previously we used recursive struct definitions to maintain the search tree.
This had problems with detecting that the root was its own parent, so we
replaced the recursive pointer at the root with an option type for all nodes.

Now a node is a root iff it has None for a parent.